requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pypdfium2>=4.0.0
paddlepaddle>=2.5.0
paddleocr>=2.7.0
python-dotenv>=1.0.0
requests>=2.31.0
google-genai>=0.8.0
google-generativeai>=0.8.0
pytesseract>=0.3.10
openai>=1.40.0
